Ensure script execution WorkingDirectory matches current script path
There seems to be a timing issue where the opened document is null while
the initial execution host parameters are created. This might not be the
best way, but making the `WorkingDirectory` writable and ensuring it
matches the current document WorkingDirectory seems to fix the issue
around relative path file loading.

Fix #350